Fix ExternalRuntime.which to behave as expected.
In current master, the which command returns whichever command comes
first in the array regardless of it's presense in the path. This happens
because Array#find does not return the value of the first truthy block
but the value of the array member which yielded it. Because the result
of a failed `which` command returns a text error, this string causes
result to evaluate as true, thus making the block truthy in the wrong
situation.