Allows NODEJS_RUNTIME to give path to node runtime
Accepts an Environment Variable "NODEJS_RUNTIME" containing the path
to the node command on systems with alternate directory structures.

    NODEJS_RUNTIME=/usr/local/bin/node

The variable is not required, and ExecJS will look for the executable
names in the PATH.

Under FreeBSD, Passenger + Rails fails to load the node runtime, even
though it is installed at /usr/local/bin/node, and control of the PATH
variable is not available.